improve vvtest -i output a little
When using "vvtest -i", you can include filtering options.  For example,
"vvtest -i -K pass" will only show tests that did not pass.  Before this
commit, the output showed skipped tests from previous runs but also the
tests that were skipped because of "-K pass" (the filtering on the
vvtest -i command line).  This is slightly confusing, so now the tests
being filtered out due to the vvtest -i command line are NOT included
in the summary information at all.