add Ctrl-C handling in batch mode
A Ctrl-C is now caught (actually any shutdown or exception) and pending
or running batch jobs are cancelled.  The batch queue implementation
has to implement the cancel(jobid) function for this to take effect,
and this was implemented for SLURM.

Also, a little more refactoring in the batching code.

Also, catch KeyboardInterrupt at vvtest entry point and don't print
a traceback.

this closes issue #2