change default Google repo group name to "code"
With Google repo manifests, there are a set of "groups" which contain
one or more "projects" (Git URLs).  Each group is defined by a file
called <name>.xml, including a file called default.xml, which is the
default group.

For mrgit clone, the name of the group is used as name of the checkout
directory.  Having the default name called "default" is undesireable,
so this commit uses the name "code" instead.