dont warn about invalid XML if not a test file
During test scanning, all files with extension 'xml' are examined
(by default).  If the xml does not start with an "<rtest" element,
then the file is ignored.  However, if the file contains invalid
XML, then a warning is printed by vvtest and then ignored.

This commit uses a simple heuristic that is applied upon an XML
error to determine if the file looks like a vvtest file.  If it
does not seem to be a vvtest file, no warning is printed.

The goal is to avoid warnings for files that are not test files.