add more info to results output file
the results files written to the TestResults.* directory only
contained 'cmdline', now they contain the same values that are
set and available in the results writer classes, including cmdline.

this is to support running "vvtest -i --cdash" which should pick
up all required information from a previous execution and submit
it to cdash.

still need to read in and use the previous information before
calling down to the cdash writer