fix file name too long bug
When vvtest cleans out the TestResults directory, it uses a "fault
tolerant" algorithm where if a path cannot be deleted, it renames it to
"remove_*_<random>" and moves on.  However, if that path fails to be
removed the next time around too, it renames it to
"remove_remove_*_<random>_<random>".  If the file removes keep failing,
the file name gets too long.

This has been "fixed" by only renaming the directory the first time.