do not apply default max procs filter in restart mode
changing into the TestResults.* directory is called "restart mode"

by default, vvtest probes for the number of processors on the machine
and filters out tests that require more than that max.  this can be
overridden with the -N option

this commit changes the behavior so that the default max procs
filter is not applied in restart mode.  however, using -N is still
respected

this also fixes a regression, where if --include-tdd was given
to start with, in restart mode it would be "forgotten"