refactoring: isolate where the test size is calculated
the test size has always been 'np' but now is 'np' and 'ndevice'.
this adds a function to TestCase to get the size of the test, which
is a pair (num procs, num devices).  And all the places where the
'np' value was calculated have been changed to use this size function