use better oversubscription algorithm
This algorithm always returns a unique set of hardware resource ids.
As before, the number of ids is always the number requested.
The "maximum available" is equivalent to the "number of hardware ids".
If a test requests more than the max available, then some of the ids
will repeat.  This should not be a problem, however, because the ids
are not used by tests for CPU core management, and the max filtering
will exclude tests taking more devices than the max available.  That
is, using --max-devices 4 on the command line will natually exclude
all tests taking more than 4 devices.

When oversubscribing, the ids given to tests will be the least loaded
ids at the time the request is made.

Also implemented lazy pooling object construction, which should be
transparent and only affects the batching use case.