add more CDash specifications
the group, site, name, and date can now be specified by including
them in the --cdash specification string, such as

  --cdash "http://sparky.com/cdash, project=Proj, group=Big Deal"

The specification must start with a file name or a URL, then follow
with name=value attributes separated by commas. The attributes are
project, group, site, name, and date. The "date" must be seconds
since epoch.