remote python: now sending a module object just makes it available
This is a behavior change for code using pythonproxy.py or
remotepython.py. A module object can be sent to the remote side, but
before this commit, the code itself in the module was executed in the
remote (global) namespace. After this commit, the module is placed in
the sys.modules dictionary on the remote (and executed within the module
namespace). This means that now remote code will have to import the
module before using it.

Also in this commit is some refactoring of filecopy.py and filesync.py
above and beyond just making them work again after the above change.