add run time skip capability
A test can now exit with a special exit status that marks the test as
a skip. The target use case is when a test cannot run due to the lack
of a resource or product on the platform. If deciding such a thing
requires some probing, then it must be done at run time. In a test,
you could do this

        import vvtest_util as vvt
        sys.exit( vvt.skip_exit_status )

Issue #70