refactoring: remove default "gcc" value for compiler
The vvtest platform object always defined non-empty value for the
compiler. If the platform initialization did not define it, then it
defaulted to "gcc". It now will not default to a value.

It is used mainly in informational ways, but also in the runtimes
caching and handling code. That is, I don't think current users of
vvtest will see an effect.