refactoring: simplify test ID mechanism
The feature added a while ago that shortens long execution directory names
(to avoid filename too long errors) meant a more intrusive test ID mechanism.
The test ID tuple as well as some derived file paths are sensitive to the
shortened directory names option.

This was done by storing the short option value in an ID generator class and
the class was carried around whenever the ID was needed, such as constructing
a new TestSpec or when changing the ID.

This commit replaces the ID generator class mechanism with a simple dictionary.
This was possible because the data is now being stored in the test list files.