adjust default ssh options for remote python execution
An ssh session consists of the client on the local side and the server on the
remote side. If the local/client side is interrupted/killed, then commands or
subprocesses running on the remote side are interrupted/killed only if the
remote ssh session is interactive (has a tty).

That is, if the command "ssh sparky sleep 60" is executed and followed a few
seconds later by ctrl-C, then the "sleep 60" command will NOT be killed.
This is because by default, ssh with a command to run will not create a tty
on the remote side.

If the command "ssh -t -t sparky sleep 60" is executed and ctrl-C is hit,
then the "sleep 60" will be killed. This is because signals are propogated
if a tty exists, and the -t -t options (both may be needed) force ssh to
create a tty on the remote side.