fix remote python orphan issue
Using ssh to a machine and executing python will place the python
process in its own process group. But this means when the local
process dies (from Ctrl-C for example), the remote python never
receives a signal and it becomes orphaned.

A previous commit used "ssh -t -t" to avoid having ssh place the
remote process in its own process group, but it turns out to have
a fatal side effect - it breaks the communication tunneling.

This commit takes a different approach. It uses a thread in the
remote python coding to detect when it has become orphaned. If
so, it sends a SIGTERM signal to its process group.