add parameter variable substitution in "depends on" patterns
Consider a test file "mytest.vvt" with these directives:

  #VVT: parameterize : toy = ball doll
  #VVT: depends on : name*toy=${toy}*

After this commit, the "${toy}" construct in the "depends on " directive
would be replaced by the value of the toy parameter. So for the test
mytest.toy=ball, the depends on pattern would actually be name*toy=ball.
Similarly, for mytest.toy=doll, the pattern would be name*toy=doll.

The use case for this feature is to connect up dependency tests to like
parameter name/values.