use a TestCase factory object to abstract away the details
To support an 'nnodes' parameter in test specifications, the cores per
node (or ppn) must be known by TestCase objects (at certain points).
But ppn is a platform/batch system thing. So instead of passing around
platform knowledge to the locations where TestCase's are created, this
sends down a factory object. The factory object can then cache platform
information (or anything really) unbeknownst to the places needing new
TestCase objects.

For this commit, a super simple factory class is defined and used, but
that will change in a following commit.