separate the Platform class from the plugin interface
The config/idplatform.py and config/platform_plugin.py plugin mechanism
is an interface from the user/project to the internal implementation.
This commit creates a PlatformConfig class that mimics a Platform object
that is being sent into the plugin functions. This means the plugin
interfacing no longer depends on the Platform class itself, and is now
free to change without affecting the interface. Similarly, the interface
can be evolved without changing the Platform class. Software design is
hard!