remove batch "shim" concept and cleanup the batch files
There was a mechanism to insert csh code into the batch files that
vvtest submits to the batch queue. This was called a shim. However,
it is not currently being used (that I know of) and all batch schedulers
will either forward the environment to the batch job automatically,
or you can add an option to have it do that. So there is no reason for
the shim.

Doing this allows the batch script to be written in bash rather than
csh (an historic thing).