using parameter nnode now works in batch and direct modes
If the parameter 'nnode' is given in a test, such as

    #VVT: parameterize : nnode = 1 2 4

then the test will consume that number of nodes times the number
of cores per compute node. Similarly for devices (GPUs). On a
workstation, a "compute node" is considered the whole workstation
cores_per_node (ppn) is specified on the command line or in the
platform plugin file.

In the case where 'nnode' is specified but not 'np', then the 'np'
parameter will still be placed in the vvtest_util.py file for
convenience. Similarly, 'ndevice' will be added for platforms that
have devices.